refactor(get): extract get
